packaging: add helper script buildsome-deps RPM
The packaging/make-deps-rpm.sh script can be used to create an
buildsome-deps RPM out of a freshly created cabal sandbox that
contains only the dependencies of buildsome.

The buildsome-deps RPM once installed, puts a ready-made cabal
sandbox under /usr/share/buildsome-deps, that buildsome's own
spec file can use.

Note that the resulting cabal builds are not built in a
distribution sandbox (a la mock), and mock is only used for
packaging in this case. Thus, the target of the build can
only be the current distribution level.

This makes it easier to rebuild buildsome itself in detached
environments (for example, under Fedora's mock, or in CentOS),
as long as the dependencies needed to build buildsome don't
change.

  ./packaging/make-deps-rpm.sh

Eventually, the result file is moved to the current directory:

   buildsome-deps-0.3.0-1.el7.centos.x86_64.rpm

Prerequisites to running ./packaging/make-deps-rpm.sh are:

  * mock installed (yum install mock)
  * 'cabal' exec in PATH that support sandboxes feature
  * a writable and empty /usr/share/buildsome-deps