Use shared memory write for no-delay notification instead of socket
After a new connection, a shared memory unlinked file is created
by the FSHooks and its read-only FD is sent over to Buildsome.

When the FSHooks see that it can append to that shared memory, it
does so instead of sending over the UNIX socket. When the connection
is closed, no more writes are assumed, and then Buildsome reads
the accumulated writes using an interator provided by the C code.